refactor DashCardActionButtons folder -> DashCardActionsPanel
