+++ frontend/src/metabase/dashboard/components/DashCard/DashCardActionsPanel/ActionSettingsButton.tsx
