+++ frontend/src/metabase/dashboard/components/DashCard/DashCardTabMenu/DashCardTabMenu.styled.tsx
@@ -1,5 +1,5 @@
 import styled from "@emotion/styled";
-import DashCardActionButton from "../DashCardActionButtons/DashCardActionButton";
+import DashCardActionButton from "../DashCardActionsPanel/DashCardActionButton";
 
 /**
  * This is a button with a bigger "hover area" to make it easier
